Split navigation building out of OpenApiGenerator
OpenApiGenerator (732 lines at the start of this branch) is now a thin
134-line orchestrator: read specs, build navigation, render pages to
disk. ApiNavigationBuilder owns the tree assembly (classifications,
tags, endpoint grouping, schema type pages, intro/outro markdown) and
the domain contracts move to their own files (IApiModel,
ApiClassification, ApiTag, ApiEndpoint).

CreateNavigation remains available on OpenApiGenerator as a forwarder
so existing callers and tests are unaffected.
