+++ src/Elastic.ApiExplorer/ApiEndpoint.cs
@@ -1,0 +1,14 @@
+// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+using System.IO.Abstractions;
+using Elastic.ApiExplorer.Operations;
+
+namespace Elastic.ApiExplorer;
+
+public record ApiEndpoint(List<ApiOperation> Operations, string? Name) : IApiGroupingModel
+{
+	/// <inheritdoc />
+	public Task RenderAsync(FileSystemStream stream, ApiRenderContext context, CancellationToken ctx = default) => Task.CompletedTask;
+}
