+++ src/Elastic.ApiExplorer/IApiModel.cs
@@ -1,0 +1,12 @@
+// Licensed to Elasticsearch B.V under one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+using Elastic.Documentation.Navigation;
+using Elastic.Documentation.Site.Navigation;
+
+namespace Elastic.ApiExplorer;
+
+public interface IApiModel : INavigationModel, IPageRenderer<ApiRenderContext>;
+
+public interface IApiGroupingModel : IApiModel;
